add AES-GCM symmetric encryption

+++ lora.h
@@ -1,3 +1,3 @@
 int lora_init(forward_data_cb_t *forwarder);
-size_t lora_send(char *msg, size_t len);
+int lora_send(char *msg, size_t len);
 void lora_listen(void);
